refactor: properties 위치 변경
