+++ backend/src/main/java/middle_point_search/backend/common/properties/JwtProperties.java
@@ -1,4 +1,4 @@
-package middle_point_search.backend.common.security.properties;
+package middle_point_search.backend.common.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
